chore: refactored MediaProgressBarComponent to ProgressBar
