+++ src/components/media-player/media-player-side.component.tsx
@@ -9,7 +9,7 @@ import { MediaPlayerService } from '../../services';
 
 import { Icon } from '../icon/icon.component';
 import { Button } from '../button/button.component';
-import { MediaProgressBarComponent } from '../media-progress-bar/media-progress-bar.component';
+import { ProgressBar } from '../progress-bar/progress-bar.component';
 import { RouterLinkToggle } from '../router-link-toggle/router-link-toggle.component';
 
 import styles from './media-player.component.css';
@@ -84,7 +84,7 @@ export function MediaPlayerSide() {
           <Icon name={mediaVolumeButtonIcon}/>
         </Button>
         <div className={cx('media-player-volume-bar-container')}>
-          <MediaProgressBarComponent
+          <ProgressBar
             autoCommitOnUpdate
             value={mediaPlaybackVolumeMuted
               ? 0
