+++ src/components/progress-bar/progress-bar.component.scss
